feat: 반려동물 정보 수정 기능 (#43)
* move: PetType 파일 위치 변경

* move: Gender 파일 위치 변경

* chore: 반려동물 몸무게 형식 double로 변경

* feat: 반려동물 정보 수정 api 생성

* chore: 반려동물 몸무게 자료형 변경에 따른 mapper 수정

+++ src/main/java/com/kau/capstone/domain/pet/util/PetMapper.java
@@ -3,8 +3,8 @@ package com.kau.capstone.domain.pet.util;
 import com.kau.capstone.domain.pet.dto.request.PetRegistRequest;
 import com.kau.capstone.domain.pet.dto.response.PetInfoResponse;
 import com.kau.capstone.domain.pet.entity.Pet;
-import com.kau.capstone.global.common.Gender;
-import com.kau.capstone.global.common.PetType;
+import com.kau.capstone.domain.pet.entity.Gender;
+import com.kau.capstone.domain.pet.entity.PetType;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
@@ -31,7 +31,7 @@ public class PetMapper {
             .petType(pet.getPetType())
             .gender(pet.getGender())
             .name(pet.getName())
-            .weight(pet.getWeight())
+            .weight((float) pet.getWeight())
             .imageUrl(pet.getImageUrl())
             .build();
     }
